requirements.txt
# Core dependencies
autogen-agentchat
autogen-ext[openai,azure]
beautifulsoup4
docxtpl
flaml[automl]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pydantic>=2.0.0
streamlit
openpyxl
openai>=1.12.0
pandas
Pillow
PyMuPDF
python-docx
streamlit-option-menu
requests
playwright
lxml
google-generativeai
pypdf2
pymupdf
gspread
rapidfuzz
oauth2client
streamlit-modal
weasyprint